{"schema":"decodex/commit/1","summary":"Modularize status and store dispatch surfaces","authority":"manual"}
