+++ apps/decodex/src/agent/tracker_tool_bridge/tools/dispatch/transition.rs
@@ -1,0 +1,63 @@
+use serde_json::Value;
+
+use crate::agent::tracker_tool_bridge::{
+	DynamicToolCallResponse, ISSUE_REVIEW_HANDOFF_TOOL_NAME, TrackerToolBridge, TransitionArgs,
+};
+
+impl<'a> TrackerToolBridge<'a> {
+	pub(super) fn handle_transition(&self, arguments: Value) -> DynamicToolCallResponse {
+		let parsed = match serde_json::from_value::<TransitionArgs>(arguments) {
+			Ok(parsed) => parsed,
+			Err(error) => {
+				return DynamicToolCallResponse::failure(format!(
+					"Invalid `issue.transition` arguments: {error}"
+				));
+			},
+		};
+
+		if let Err(error) = self.ensure_issue_scope(&parsed.scope) {
+			return DynamicToolCallResponse::failure(error);
+		}
+
+		let allowed_states = self.allowed_transition_states();
+
+		if !allowed_states.iter().any(|state| state == &parsed.state) {
+			let success_state = self.workflow.frontmatter().tracker().success_state();
+
+			if parsed.state == success_state {
+				return DynamicToolCallResponse::failure(format!(
+					"State `{}` requires `{}` after the branch is pushed and a reviewable PR exists.",
+					parsed.state, ISSUE_REVIEW_HANDOFF_TOOL_NAME
+				));
+			}
+
+			return DynamicToolCallResponse::failure(format!(
+				"State `{}` is outside the allowed tracker tool policy.",
+				parsed.state
+			));
+		}
+
+		let Some(state_id) = self.issue.state_id_for_name(&parsed.state) else {
+			return DynamicToolCallResponse::failure(format!(
+				"State `{}` does not exist on issue `{}`.",
+				parsed.state, self.issue.identifier
+			));
+		};
+
+		match self.tracker.update_issue_state(&self.issue.id, state_id) {
+			Ok(()) => {
+				self.local_issue_state_name.replace(parsed.state.clone());
+				self.record_continuation_blocking_transition(&parsed.state);
+
+				DynamicToolCallResponse::success(format!(
+					"Issue `{}` moved to `{}`.",
+					self.issue.identifier, parsed.state
+				))
+			},
+			Err(error) => DynamicToolCallResponse::failure(format!(
+				"Failed to move issue `{}` to `{}`: {error}",
+				self.issue.identifier, parsed.state
+			)),
+		}
+	}
+}
