+++ apps/decodex/src/state/store/decision_contracts.rs
@@ -1,14 +1,10 @@
+mod mutation;
+mod query;
+
 use crate::{
 	loop_contract::{DecisionContract, DecisionPromotion},
-	prelude::{Result, eyre},
-	state::{
-		runtime_records::{DecisionContractKey, DecisionContractRuntimeRecord},
-		store,
-		store::{
-			DecisionContractRecord, StateStore, compare_decision_contract_runtime_records,
-			validation,
-		},
-	},
+	prelude::Result,
+	state::store::{DecisionContractRecord, StateStore},
 };
 
 impl StateStore {
@@ -18,34 +14,7 @@ impl StateStore {
 		source_issue_id: Option<&str>,
 		contract: DecisionContract,
 	) -> Result<DecisionContractRecord> {
-		validation::validate_decision_contract_record_inputs(
-			project_id,
-			source_issue_id,
-			&contract,
-		)?;
-
-		let now = store::timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-		let key = DecisionContractKey::new(project_id, contract.contract_id());
-		let (created_at, created_at_unix) = state.decision_contracts.get(&key).map_or_else(
-			|| (now.text.clone(), now.unix),
-			|record| (record.created_at.clone(), record.created_at_unix),
-		);
-		let record = DecisionContractRuntimeRecord {
-			project_id: project_id.to_owned(),
-			source_issue_id: source_issue_id.map(str::to_owned),
-			status: contract.status(),
-			contract,
-			created_at,
-			created_at_unix,
-			updated_at: now.text,
-			updated_at_unix: now.unix,
-		};
-
-		state.decision_contracts.insert(record.key(), record.clone());
-		self.upsert_decision_contract_locked(&record)?;
-
-		Ok(record.as_public())
+		mutation::upsert_decision_contract(self, project_id, source_issue_id, contract)
 	}
 
 	/// Read one local Loop/Decision Contract by project and contract id.
@@ -55,23 +24,7 @@ impl StateStore {
 		project_id: &str,
 		contract_id: &str,
 	) -> Result<Option<DecisionContractRecord>> {
-		validation::validate_required_decision_contract_field("project_id", project_id)?;
-		validation::validate_required_decision_contract_field("contract_id", contract_id)?;
-
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-
-			return sqlite
-				.decision_contract(project_id, contract_id)
-				.map(|record| record.map(|record| record.as_public()));
-		}
-
-		let state = self.lock()?;
-
-		Ok(state
-			.decision_contracts
-			.get(&DecisionContractKey::new(project_id, contract_id))
-			.map(DecisionContractRuntimeRecord::as_public))
+		query::decision_contract(self, project_id, contract_id)
 	}
 
 	/// List local Loop/Decision Contracts sourced from one tracker issue.
@@ -81,34 +34,7 @@ impl StateStore {
 		project_id: &str,
 		source_issue_id: &str,
 	) -> Result<Vec<DecisionContractRecord>> {
-		validation::validate_required_decision_contract_field("project_id", project_id)?;
-		validation::validate_required_decision_contract_field("source_issue_id", source_issue_id)?;
-
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-			let records = sqlite
-				.list_decision_contracts_for_issue(project_id, source_issue_id)?
-				.into_iter()
-				.map(|record| record.as_public())
-				.collect();
-
-			return Ok(records);
-		}
-
-		let state = self.lock()?;
-		let mut records = state
-			.decision_contracts
-			.values()
-			.filter(|record| {
-				record.project_id == project_id
-					&& record.source_issue_id.as_deref() == Some(source_issue_id)
-			})
-			.cloned()
-			.collect::<Vec<_>>();
-
-		records.sort_by(compare_decision_contract_runtime_records);
-
-		Ok(records.into_iter().map(|record| record.as_public()).collect())
+		query::list_decision_contracts_for_issue(self, project_id, source_issue_id)
 	}
 
 	/// List local Loop/Decision Contracts for one project.
@@ -117,30 +43,7 @@ impl StateStore {
 		&self,
 		project_id: &str,
 	) -> Result<Vec<DecisionContractRecord>> {
-		validation::validate_required_decision_contract_field("project_id", project_id)?;
-
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-			let records = sqlite
-				.list_decision_contracts_for_project(project_id)?
-				.into_iter()
-				.map(|record| record.as_public())
-				.collect();
-
-			return Ok(records);
-		}
-
-		let state = self.lock()?;
-		let mut records = state
-			.decision_contracts
-			.values()
-			.filter(|record| record.project_id == project_id)
-			.cloned()
-			.collect::<Vec<_>>();
-
-		records.sort_by(compare_decision_contract_runtime_records);
-
-		Ok(records.into_iter().map(|record| record.as_public()).collect())
+		query::list_decision_contracts_for_project(self, project_id)
 	}
 
 	/// Promote a latent Loop/Decision Contract into accepted execution authority.
@@ -189,29 +92,6 @@ impl StateStore {
 		contract_id: &str,
 		update: impl FnOnce(&mut DecisionContract) -> Result<()>,
 	) -> Result<DecisionContractRecord> {
-		validation::validate_required_decision_contract_field("project_id", project_id)?;
-		validation::validate_required_decision_contract_field("contract_id", contract_id)?;
-
-		let now = store::timestamp_parts();
-		let key = DecisionContractKey::new(project_id, contract_id);
-		let mut state = self.lock()?;
-		let mut record = state
-			.decision_contracts
-			.get(&key)
-			.cloned()
-			.ok_or_else(|| eyre::eyre!("Decision contract `{contract_id}` does not exist."))?;
-
-		update(&mut record.contract)?;
-
-		record.contract.validate()?;
-
-		record.status = record.contract.status();
-		record.updated_at = now.text;
-		record.updated_at_unix = now.unix;
-
-		state.decision_contracts.insert(key, record.clone());
-		self.upsert_decision_contract_locked(&record)?;
-
-		Ok(record.as_public())
+		mutation::update_decision_contract(self, project_id, contract_id, update)
 	}
 }
